Fixed; add models.py
