+++ bot.py
@@ -39,12 +39,10 @@ def load_responses_from_files(file: Path) -> str:
             return result
     except FileNotFoundError as file_not_found_err:
         logging.error(f"Err while load file - {file_not_found_err}")
-        return ""
+        return None
 
 
-async def send_message_template(
-    user_data, event, start_range, end_range
-):
+async def send_message_template(user_data, event, start_range, end_range):
     logging.info(
         f"Contact: {user_data.contact} -"
         f"username: {user_data.first_name} - "
@@ -61,34 +59,10 @@ Hello, {user_data.first_name}. \n
 **This message was sent automatically.** \n
 """,
         )
-        await CLIENT.send_message(user_data.id, )
+        await CLIENT.send_message(
+            user_data.id,
+        )
         logging.info(f"Response was sent to {user_data.first_name}.")
-
-
-# @CLIENT.on(events.NewMessage)
-# async def response_to_group(event):
-#
-#     user_data = await event.client.get_entity(event.from_id)
-#
-#     logging.info(f"Raw sender data: {user_data}")
-#
-#     try:
-#         if user_data.id in 123:
-#             await send_message_template(user_data, event, 5, 10, )
-#         elif user_data.id in 123:
-#             await send_message_template(user_data, event, 15, 20, )
-#         elif not user_data.contact:
-#             for key_words in 123:
-#                 if key_words in event.message.message:
-#                     logging.info("Looks like HR is on the line.")
-#                     await send_message_template(user_data, event, 1, 5, )
-#     except ValueError as val_err:
-#         logging.error(f"Sender is {user_data.first_name}")
-#         logging.error(val_err)
-#     except TypeError as type_err:
-#         logging.error("That maybe sticker was sent, not text.")
-#         logging.error(f"Sender is {user_data.first_name}")
-#         logging.error(type_err)
 
 
 if __name__ == "__main__":
